Attach - Notice mapping

+++ src/main/java/bleuauction/bleuauction_be/server/notice/controller/NoticeController.java
@@ -1,12 +1,13 @@
 package bleuauction.bleuauction_be.server.notice.controller;
 
+import bleuauction.bleuauction_be.server.attach.entity.Attach;
+import bleuauction.bleuauction_be.server.attach.service.AttachService;
 import bleuauction.bleuauction_be.server.member.entity.Member;
+import bleuauction.bleuauction_be.server.ncp.NcpObjectStorageService;
 import bleuauction.bleuauction_be.server.notice.entity.Notice;
-import bleuauction.bleuauction_be.server.notice.entity.NoticeStatus;
 import bleuauction.bleuauction_be.server.notice.service.NoticeService;
 import bleuauction.bleuauction_be.server.notice.web.NoticeForm;
 import jakarta.persistence.EntityManager;
-import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
@@ -14,6 +15,7 @@ import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.multipart.MultipartFile;
 
 
 import java.util.ArrayList;
@@ -26,6 +28,8 @@ public class NoticeController {
 
   private final NoticeService noticeService;
   private final EntityManager entityManager;
+  private final NcpObjectStorageService ncpObjectStorageService;
+  private final AttachService attachService;
 
 
   //등록
@@ -39,19 +43,26 @@ public class NoticeController {
   // 등록 처리
   @PostMapping("/api/notice/new")
   @Transactional
-  public ResponseEntity<String>  notice(@Valid @RequestBody NoticeForm form) {
-    Notice notice = new Notice();
+  public ResponseEntity<String>  notice(Notice notice, @RequestParam(name = "multipartFiles",required = false) List<MultipartFile> multipartFiles) {
     Member member =entityManager.find(Member.class, 1L);
-    notice.setNoticeTitle(form.getNoticeTitle());
-    notice.setNoticeContent(form.getNoticeContent());
-    notice.setMember(member);
-    notice.setNoticeStatus(NoticeStatus.Y);
+    notice.setMemberNo(member);
+
+    if (multipartFiles != null && multipartFiles.size() > 0) {
+      ArrayList<Attach> attaches = new ArrayList<>();
+      for (MultipartFile multipartFile : multipartFiles) {
+        if (multipartFile.getSize() > 0) {
+          Attach attach = ncpObjectStorageService.uploadFile(new Attach(),
+                  "bleuauction-bucket", "notice/", multipartFile);
+          notice.addNoticeAttach(attach);
+        }
+      }
+    }
+
     notice = entityManager.merge(notice);
-    noticeService.enroll(notice); // 서비스 메소드에서 엔터티 등록
+    noticeService.enroll(notice);
 
     log.info("notice/postnew");
 
-    //return "redirect:/notices";
     return ResponseEntity.status(HttpStatus.CREATED).body("Notice created successfully");
   }
 
@@ -70,18 +81,43 @@ public class NoticeController {
   }
 
 
-// 삭제 처리
+// 삭제
   @PostMapping("/api/notice/delete/{noticeNo}")
   public ResponseEntity<String> deleteNotice(@PathVariable("noticeNo") Long noticeNo) {
-    noticeService.deleteNotice(noticeNo);
-    return ResponseEntity.ok("Notice deleted successfully");
+    Notice notice = noticeService.findOne(noticeNo);
+
+    // 사진 상태를 'N'으로 변경
+    if (notice != null) {
+      if (notice.getNoticeAttaches() != null && !notice.getNoticeAttaches().isEmpty()) {
+        for (Attach attach : notice.getNoticeAttaches()) {
+          attachService.update(attach.getFileNo());
+        }
+      }
+
+      noticeService.deleteNotice(noticeNo);
+      return ResponseEntity.ok("Notice deleted successfully");
+    } else {
+      return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Notice not found");
+
+    }
   }
+
+  //사진삭제 -- attach 컨트롤러에 모으기?
+  @DeleteMapping("/api/notice/deletefile/{fileNo}")
+  public ResponseEntity<String> fileNoticeDelete(@PathVariable Long fileNo) {
+    attachService.update(fileNo);
+    return ResponseEntity.ok("File deleted successfully");
+  }
+
+
 
   //디테일(수정)
   @GetMapping("/api/notice/detail/{noticeNo}")
   public ResponseEntity<Notice> detailNotice(@PathVariable("noticeNo") Long noticeNo) {
     Notice notice = noticeService.findOne(noticeNo);
+
     if (notice != null) {
+      notice.setNoticeAttaches(notice.getNoticeAttaches());
       return ResponseEntity.ok(notice);
     } else {
       return ResponseEntity.notFound().build();
@@ -90,16 +126,27 @@ public class NoticeController {
 
   // 수정 처리
   @PostMapping("/api/notice/update/{noticeNo}")
-  public ResponseEntity<String> updateNotice(
+  public ResponseEntity<String> updateNotice(Notice notice,
           @PathVariable("noticeNo") Long noticeNo,
-          @RequestBody NoticeForm form
-  ) {
-    Notice updated = noticeService.update(noticeNo, form.getNoticeTitle(), form.getNoticeContent());
-    if (updated!= null) {
-      return ResponseEntity.ok("Notice updated successfully");
-    } else {
-      return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Notice not found or update failed");
+          @RequestParam(name = "multipartFiles",required = false) List<MultipartFile> multipartFiles) {
+
+    Notice updatedNotice = noticeService.findOne(noticeNo);
+
+    if (multipartFiles != null && multipartFiles.size() > 0) {
+      ArrayList<Attach> attaches = new ArrayList<>();
+      for (MultipartFile multipartFile : multipartFiles) {
+        if (multipartFile.getSize() > 0) {
+          Attach attach = ncpObjectStorageService.uploadFile(new Attach(),
+                  "bleuauction-bucket", "notice/", multipartFile);
+          updatedNotice.addNoticeAttach(attach);
+        }
+      }
     }
+
+    noticeService.update(notice);
+    log.info("notice/update");
+    return ResponseEntity.ok("Notice updated successfully");
+
   }
 
 }
